class-14/pyproject.toml
[project]
name = "class-14"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.2.6",
    "python-dotenv>=1.1.1",
]
